+++ tag/cli/__init__.py
@@ -11,6 +11,7 @@ import argparse
 import sys
 import tag
 from . import bae
+from . import bcollapse
 from . import gff3
 from . import locuspocus
 from . import merge
@@ -20,6 +21,7 @@ from . import sum
 
 subparser_funcs = {
     'bae': bae.subparser,
+    'bcollapse': bcollapse.subparser,
     'gff3': gff3.subparser,
     'locuspocus': locuspocus.subparser,
     'merge': merge.subparser,
@@ -30,6 +32,7 @@ subparser_funcs = {
 
 mains = {
     'bae': bae.main,
+    'bcollapse': bcollapse.main,
     'gff3': gff3.main,
     'locuspocus': locuspocus.main,
     'merge': merge.main,